refactor: split drive auth support runtime
Drive auth path resolution remains at the root while prompt adapters and token/auth runtime helpers move into a narrower support module.
